improve state stack related game loop
