+++ package.json
@@ -20,7 +20,7 @@
   },
   "homepage": "https://github.com/w0lf3n/broken-sword#readme",
   "devDependencies": {
-    "eslint": "^8.20.0",
+    "eslint": "^8.21.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-security": "^1.5.0"
