+++ src/index.js
@@ -1,21 +1,17 @@
 
+import { loadJSON } from "../lib/JST/resource/loaders.js";
 import * as Tick from "../lib/JST/native/tick.js";
 import Application from "../lib/JST/dom/application.js";
+import * as Lang from "../lib/JST/resource/lang.js";
 //import { getDefaultRendererContext } from "../lib/JST/";
 
 import GameStates from "./game_states.js";
 import * as SetupState from "./setup_state.js";
 
 
-const GAME_TITLE = "The Broken Mirror";
-const app = new Application(GAME_TITLE);
-const ctx = app.getRootPane();
+let ctx = null;
 
-// push initial state(s) to gameStates
-GameStates.push(SetupState);
-
-// game loop
-Tick.start((dt) => {
+const game_loop = (dt) => {
 
     GameStates.update(dt);
     GameStates.render(ctx);
@@ -24,4 +20,31 @@ Tick.start((dt) => {
         Tick.stop();
         throw new Error("The game states stack is empty.");
     }
-});
+};
+
+const create_application = () => {
+
+    const GAME_TITLE = Lang.getWord("GameTitle");
+
+    Application.setTitle(GAME_TITLE);
+    Application.setContextMenuEnabled(false);
+
+    const app = new Application(GAME_TITLE);
+    app.addClass("Center");
+    ctx = app.getRootPane(); //getDefaultRendererContext();
+
+    // push initial state(s) to gameStates
+    GameStates.push(SetupState);
+
+    Tick.start(game_loop);
+};
+
+// load settings .then
+loadJSON("dat/locale/base_en.json")
+    .then((data) => {
+
+        Lang.setCurrentLanguage(Lang.Language.en);
+        Lang.addLanguagePack(data);
+
+        create_application();
+    });
