+++ src/ui/ui.js
@@ -419,97 +419,6 @@ export const Overlay = (() => {
     return Object.freeze(core);
 })();
 
-export const Menu = (() => {
-
-    let container = null;
-
-    let title = null;
-    let subtitle = null;
-
-    let buttons = null;
-
-    let disclaimer = null;
-
-    const core = {};
-
-    core.init = () => {
-
-        container = new Container("Menu");
-        container.addClass("Center");
-
-        title = new TextComponent(null, "Title");
-        subtitle = new TextComponent(null, "SubTitle");
-
-        const titleBox = new Container("TitleBox");
-        titleBox.append(title, subtitle);
-
-        buttons = new Container("ButtonBox");
-
-        disclaimer = new TextComponent(null, "Disclaimer");
-
-        container.append(titleBox, buttons, disclaimer);
-
-    };
-
-    core.setTitle = (val) => {
-        if (title instanceof TextComponent) {
-            title.text = val;
-        }
-    };
-
-    core.setSubTitle = (val) => {
-        if (subtitle instanceof TextComponent) {
-            subtitle.text = val;
-        }
-    };
-
-    core.addButton = (text, action) => {
-        if (buttons instanceof Container) {
-            buttons.addComponent(new TextButton(text, action));
-        }
-    };
-
-    core.clearButtons = () => {
-        if (buttons instanceof Container) {
-            buttons.clear();
-        }
-    };
-
-    core.setDisclaimer = (val) => {
-        if (disclaimer instanceof TextComponent) {
-            disclaimer.text = val;
-        }
-    };
-
-    core.show = () => {
-        if (container instanceof Container) {
-            container.show();
-        }
-    };
-
-    core.hide = () => {
-        if (container instanceof Container) {
-            container.hide();
-        }
-    };
-
-    core.getContainer = () => container;
-
-    return Object.freeze(core);
-})();
-
-export const setup = (title) => {
-
-    app = new Application(title);
-    app.addClass("Center");
-    Application.setTitle(title);
-    Application.setContextMenuEnabled(false);
-
-    Menu.init();
-    Menu.hide();
-    app.getRootPane().addComponent(Menu.getContainer());
-};
-
 export const initGameUI = (listener) => {
 
     Overlay.init();
